fixing read and output format
Fixes an issue where a new line character would be printed before the prompt on some shells.

Fixes output. Outputting tab-delimited output to the shell is unreliable, since tabs tend to get replaced by some arbitrary number of space characters. It now prints to the shell and writes to a file in the working dir, since the output file reliably preserves  tabs.